SD 상품 리스트추가,보류 로직 수정/ GR PO,stk 연결 흐름 수정

+++ src/main/java/com/youthcase/orderflow/sd/sdRefund/service/RefundInventoryService.java
@@ -48,7 +48,7 @@ public class RefundInventoryService {
             Lot newLot = new Lot();
             newLot.setProduct(product);
             newLot.setExpDate(expDate);
-            newLot.setQty(BigDecimal.valueOf(item.getSalesQuantity()));
+            newLot.setQty((long) item.getSalesQuantity());
             newLot.setStatus(Lot.LotStatus.RETURNED);
             newLot.setCreatedAt(OffsetDateTime.now());
             newLot.setUpdatedAt(OffsetDateTime.now());
